fragment 구조 변경
