+++ app/src/main/java/com/oneandonly/inventationblock/ui/activity/MenuActivity.kt
@@ -5,12 +5,9 @@ import android.os.Bundle
 import android.util.Log
 import androidx.databinding.DataBindingUtil
 import androidx.fragment.app.Fragment
-import androidx.fragment.app.commit
-import androidx.fragment.app.commitNow
 import com.oneandonly.inventationblock.R
 import com.oneandonly.inventationblock.databinding.ActivityMenuBinding
-import com.oneandonly.inventationblock.ui.fragment.MenuAddFragment
-import com.oneandonly.inventationblock.ui.fragment.RegisterFragment
+import com.oneandonly.inventationblock.ui.fragment.MenuFragment
 
 class MenuActivity : AppCompatActivity() {
 
@@ -23,10 +20,7 @@ class MenuActivity : AppCompatActivity() {
         binding = DataBindingUtil.setContentView(this@MenuActivity, R.layout.activity_menu)
         binding.lifecycleOwner = this@MenuActivity
 
-        changeFragment(MenuAddFragment(),"Menu")
-        binding.menuBtn.isSelected = true
-        binding.drinkBtn.isSelected = false
-
+        changeFragment(MenuFragment(),"Menu")
         setViewModel()
         uiSetting()
     }
@@ -37,65 +31,9 @@ class MenuActivity : AppCompatActivity() {
 
     private fun uiSetting() {
         toolbarSetting()
-        fragmentSetting()
     }
 
-    private fun fragmentSetting() {
-        binding.menuBtn.setOnClickListener {
-            /*fragmentManager.commit {
-                Log.d("Fragment Test","Menu click ${fragmentManager.fragments}")
-                if (fragmentManager.findFragmentByTag("Menu") != null) {
-                    val menuFragment = fragmentManager.findFragmentByTag("Menu")
-                    val bundle = Bundle()
-                    Log.d("Fragment Test","Menu not null")
-                    bundle.putString("test","menu")
-                    menuFragment?.arguments = bundle
 
-                    replace(R.id.fl_menu, menuFragment!!)
-                } else {
-                    Log.d("Fragment Test","Menu null")
-                    add(R.id.fl_menu, MenuAddFragment(), "Menu")
-                    addToBackStack(null)
-                }
-            }*/
-            for (fragment: Fragment in supportFragmentManager.fragments) {
-                if (fragment.isVisible && fragment.tag != "Menu") {
-                    changeFragment(MenuAddFragment(),"Menu")
-                    Log.d("Fragment Test","Menu Open")
-                    binding.menuBtn.isSelected = true
-                    binding.drinkBtn.isSelected = false
-                }
-            }
-
-        }
-        binding.drinkBtn.setOnClickListener {
-            /*fragmentManager.commit {
-                Log.d("Fragment Test","Drink Click ${fragmentManager.fragments}")
-                if (fragmentManager.findFragmentByTag("Drink") != null) {
-                    val menuFragment = fragmentManager.findFragmentByTag("Drink")
-                    val bundle = Bundle()
-                    Log.d("Fragment Test","Drink not null")
-                    bundle.putString("test","drink")
-                    menuFragment?.arguments = bundle
-
-                    replace(R.id.fl_menu, menuFragment!!)
-                } else {
-                    Log.d("Fragment Test","drink null")
-                    add(R.id.fl_menu, MenuAddFragment(), "Drink")
-                    addToBackStack(null)
-                }
-            }*/
-            for (fragment: Fragment in supportFragmentManager.fragments) {
-                if (fragment.isVisible && fragment.tag != "Drink") {
-                    changeFragment(MenuAddFragment(),"Drink")
-                    Log.d("Fragment Test","Drink Open")
-                    binding.menuBtn.isSelected = false
-                    binding.drinkBtn.isSelected = true
-                }
-            }
-
-        }
-    }
 
     private fun toolbarSetting() {
         binding.menuToolBar.toolBarBackBtn.setOnClickListener {
@@ -105,15 +43,19 @@ class MenuActivity : AppCompatActivity() {
     }
 
     override fun onBackPressed() {
-        super.onBackPressed()
-        finish()
+        for(fragment: Fragment in supportFragmentManager.fragments) {
+            when (fragment.tag) {
+                "Menu" -> {
+                    finish()
+                }
+                "MenuAdd" -> {
+                    changeFragment(MenuFragment(),"Menu")
+                }
+            }
+        }
     }
 
     private fun changeFragment(fragment: Fragment, tag: String) {
-        val bundle = Bundle()
-        bundle.putString("type",tag)
-        fragment.arguments = bundle
-
         val transaction = supportFragmentManager.beginTransaction()
         transaction.replace(R.id.fl_menu, fragment, tag)
         transaction.commit()
