--- app/src/main/java/com/oneandonly/inventationblock/ui/fragment/DrinkMenuListFragment.kt
@@ -1,4 +1,0 @@
-package com.oneandonly.inventationblock.ui.fragment
-
-class DrinkMenuListFragment {
-}
