--- app/src/main/java/com/oneandonly/inventationblock/ui/fragment/MainMenuListFragment.kt
@@ -1,4 +1,0 @@
-package com.oneandonly.inventationblock.ui.fragment
-
-class MainMenuListFragment {
-}
