+++ app/src/main/res/color/menu_add_btn_background_color.xml
@@ -1,0 +1,5 @@
+<?xml version="1.0" encoding="utf-8"?>
+<selector xmlns:android="http://schemas.android.com/apk/res/android">
+    <item android:color="@color/white" android:state_pressed="false"/>
+    <item android:color="@color/red" android:state_pressed="true"/>
+</selector>
