+++ app/src/main/res/drawable/bg_red_btn.xml
@@ -3,5 +3,5 @@
     <corners android:radius="5dp"/>
     <stroke android:width="1dp"
         android:color="@color/red"/>
-    <solid android:color="@color/white"/>
+    <solid android:color="@color/menu_add_btn_background_color"/>
 </shape>
