+++ app/src/main/res/layout/activity_stock_add.xml
@@ -6,7 +6,7 @@
     <data>
         <variable
             name="date"
-            type="com.oneandonly.inventationblock.ui.fragment.MenuAddFragment" />
+            type="com.oneandonly.inventationblock.ui.fragment.MenuListFragment" />
     </data>
 
     <RelativeLayout
