+++ app/src/main/res/layout/fragment_menu.xml
@@ -11,21 +11,55 @@
         android:layout_height="match_parent">
 
         <RelativeLayout
+            android:id="@+id/table"
             android:layout_width="match_parent"
-            android:layout_height="match_parent">
+            android:layout_height="wrap_content"
+            android:layout_alignParentStart="true"
+            android:layout_alignParentEnd="true"
+            android:layout_alignParentTop="true"
+            android:layout_marginTop="@dimen/layout_margin"
+            android:layout_marginStart="@dimen/layout_margin"
+            android:layout_marginEnd="@dimen/layout_margin">
 
-            <androidx.recyclerview.widget.RecyclerView
-                android:id="@+id/menu_list"
-                android:layout_width="match_parent"
-                android:layout_height="match_parent"
+            <TextView
+                android:id="@+id/menu_btn"
+                style="@style/MenuTextStyle"
+                android:backgroundTint="@color/button_background_color"
                 android:layout_alignParentStart="true"
+                android:text="주메뉴"
+                android:textColor="@color/text_selected"
+                android:textStyle="bold"/>
+
+            <TextView
+                android:id="@+id/drink_btn"
+                style="@style/MenuTextStyle"
+                android:backgroundTint="@color/button_background_color"
+                android:layout_toEndOf="@id/menu_btn"
+                android:layout_marginStart="@dimen/layout_between_margin"
+                android:text="음료"
+                android:textColor="@color/text_selected"
+                android:textStyle="bold"/>
+
+            <TextView
+                android:id="@+id/menu_add_btn"
+                style="@style/RedButtonStyle"
                 android:layout_alignParentEnd="true"
-                android:layout_alignParentTop="true"
-                tools:listitem="@layout/item_menu_add">
-
-            </androidx.recyclerview.widget.RecyclerView>
-
+                android:layout_marginStart="@dimen/layout_between_margin"
+                android:text="메뉴 직접 추가하기"/>
         </RelativeLayout>
+
+        <FrameLayout
+            android:id="@+id/fl_menu"
+            android:layout_width="match_parent"
+            android:layout_height="match_parent"
+            android:layout_below="@+id/table"
+            android:layout_alignParentStart="true"
+            android:layout_alignParentEnd="true"
+            android:layout_alignParentBottom="true"
+            android:layout_marginStart="@dimen/layout_margin"
+            android:layout_marginTop="@dimen/layout_margin"
+            android:layout_marginEnd="@dimen/layout_margin"
+            android:layout_marginBottom="@dimen/layout_margin"/>
 
     </RelativeLayout>
 </layout>
