+++ app/src/main/res/layout/fragment_menu_add.xml
@@ -1,6 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
-<layout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools">
+<layout xmlns:android="http://schemas.android.com/apk/res/android">
 
     <data>
 
@@ -8,7 +7,27 @@
 
     <RelativeLayout
         android:layout_width="match_parent"
-        android:layout_height="match_parent">
+        android:layout_height="match_parent"
+        android:padding="@dimen/layout_margin">
 
+        <TextView
+            android:id="@+id/comment"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:layout_alignParentStart="true"
+            android:layout_alignParentTop="true"
+            android:text="* 메뉴를 직접 추가하는 페이지입니다."/>
+
+        <TextView
+            style="@style/MenuTextStyle"
+            android:backgroundTint="@color/primary_color"
+            android:text="메뉴 이름 입력"
+            android:layout_marginTop="20dp"
+            android:layout_alignParentStart="true"
+            android:layout_below="@+id/comment"/>
+        
+        <EditText
+            android:layout_width=""
+            android:layout_height=""
     </RelativeLayout>
 </layout>
