+++ app/src/main/res/values/styles.xml
@@ -85,10 +85,8 @@
         <item name="android:paddingBottom">4dp</item>
         <item name="android:paddingStart">12dp</item>
         <item name="android:paddingEnd">12dp</item>
-        <item name="android:textColor">@color/red</item>
+        <item name="android:textColor">@color/menu_add_btn_text_color</item>
         <item name="android:gravity">center</item>
     </style>
-
-
 
 </resources>
